Split large files into smaller modules for AI-friendliness
- main.rs (2045→3 lines): CLI structs → cli/args.rs, help text →
  cli/help.rs, dispatch logic → run.rs (re-exported from lib.rs)
- find.rs: split into find/mod.rs + filter_index.rs + build.rs + sort.rs
- Shared mutation index-update helpers → commands/mutation.rs
- Knowledgebase: move completed iterations to done/, add backlog items
- All quality gates pass (498 tests, zero clippy warnings)
