+++ crates/hyalo-cli/src/commands/find/mod.rs
@@ -1,7 +1,12 @@
 #![allow(clippy::missing_errors_doc)]
+
+mod build;
+mod filter_index;
+mod sort;
+
+pub use filter_index::{filter_index_entries, needs_body};
+
 use anyhow::{Context, Result};
-use globset::{GlobBuilder, GlobSetBuilder};
-use std::collections::HashSet;
 use std::path::Path;
 
 use crate::output::{CommandOutcome, Format};
@@ -9,151 +14,14 @@ use hyalo_core::content_search::ContentSearchVisitor;
 use hyalo_core::discovery;
 use hyalo_core::filter::{self, Fields, FindTaskFilter, PropertyFilter, SortField};
 use hyalo_core::heading::{SectionFilter, SectionRange, build_section_scope, in_scope};
-use hyalo_core::index::{IndexEntry, VaultIndex};
-use hyalo_core::link_graph::{LinkGraph, is_self_link};
+use hyalo_core::index::VaultIndex;
+use hyalo_core::link_graph::is_self_link;
 use hyalo_core::types::{
     BacklinkInfo, ContentMatch, FileObject, FindTaskInfo, LinkInfo, OutlineSection, PropertyInfo,
 };
 
-// ---------------------------------------------------------------------------
-// Public command entry point
-// ---------------------------------------------------------------------------
-
-/// Filter index entries by optional `--file` and `--glob` scoping arguments.
-///
-/// - When both are empty, all entries are returned.
-/// - `files_arg` entries are matched by exact `rel_path` equality.
-/// - `globs` patterns use the same globset semantics as `discovery::match_globs`
-///   (positive patterns require a match; negative `!pat` patterns exclude matches).
-pub fn filter_index_entries<'a>(
-    entries: &'a [IndexEntry],
-    files_arg: &[String],
-    globs: &[String],
-) -> Result<Vec<&'a IndexEntry>> {
-    if files_arg.is_empty() && globs.is_empty() {
-        return Ok(entries.iter().collect());
-    }
-
-    if !files_arg.is_empty() && globs.is_empty() {
-        // Exact path matching — build a HashSet for O(1) lookup instead of O(n×m)
-        let file_set: HashSet<&str> = files_arg.iter().map(String::as_str).collect();
-        let filtered: Vec<&IndexEntry> = entries
-            .iter()
-            .filter(|e| file_set.contains(e.rel_path.as_str()))
-            .collect();
-        return Ok(filtered);
-    }
-
-    // Glob filtering — build positive/negative glob sets (same logic as discovery::match_globs)
-    let normalized: Vec<String> = globs
-        .iter()
-        .map(|p| {
-            if let Some(rest) = p.strip_prefix("\\!") {
-                format!("!{rest}")
-            } else {
-                p.clone()
-            }
-        })
-        .collect();
-
-    let mut positive: Vec<&str> = Vec::new();
-    let mut negative: Vec<&str> = Vec::new();
-    for p in &normalized {
-        if let Some(neg) = p.strip_prefix('!') {
-            anyhow::ensure!(
-                !neg.is_empty(),
-                "negation glob pattern must not be empty (got '!')"
-            );
-            negative.push(neg);
-        } else {
-            positive.push(p.as_str());
-        }
-    }
-
-    let positive_set = if positive.is_empty() {
-        None
-    } else {
-        let mut builder = GlobSetBuilder::new();
-        for pat in &positive {
-            builder.add(
-                GlobBuilder::new(pat)
-                    .literal_separator(true)
-                    .build()
-                    .context("invalid glob pattern")?,
-            );
-        }
-        Some(
-            builder
-                .build()
-                .context("failed to build positive globset")?,
-        )
-    };
-
-    let negative_set = if negative.is_empty() {
-        None
-    } else {
-        let mut builder = GlobSetBuilder::new();
-        for pat in &negative {
-            builder.add(
-                GlobBuilder::new(pat)
-                    .literal_separator(true)
-                    .build()
-                    .context("invalid glob negation pattern")?,
-            );
-        }
-        Some(
-            builder
-                .build()
-                .context("failed to build negative globset")?,
-        )
-    };
-
-    let filtered: Vec<&IndexEntry> = entries
-        .iter()
-        .filter(|e| {
-            // When files_arg is non-empty, an entry also passes if it is listed
-            // explicitly — this handles the case where both files_arg and globs
-            // are present (OR semantics between the two inclusion criteria).
-            let passes_files = !files_arg.is_empty() && files_arg.iter().any(|f| f == &e.rel_path);
-            let passes_positive = positive_set
-                .as_ref()
-                .is_none_or(|gs| gs.is_match(&e.rel_path));
-            let passes_negative = negative_set
-                .as_ref()
-                .is_none_or(|gs| !gs.is_match(&e.rel_path));
-            // An entry is included if it matches files_arg OR (matches positive
-            // globs AND is not excluded by a negative glob).
-            passes_files || (passes_positive && passes_negative)
-        })
-        .collect();
-
-    Ok(filtered)
-}
-
-/// Returns `true` when the command needs body content (sections, tasks, links,
-/// title, content search, or structural filters).
-///
-/// This is the core body-scan predicate shared between the `find` command
-/// dispatch in `main.rs` and any other caller that needs to decide whether to
-/// request a full body scan from [`crate::commands::build_scanned_index`].
-///
-/// Callers may add extra conditions on top (e.g. `sort_needs_links`,
-/// `broken_links`, `has_title_filter`) that are specific to their dispatch
-/// context.
-pub fn needs_body(
-    fields: &Fields,
-    has_content_search: bool,
-    has_task_filter: bool,
-    has_section_filter: bool,
-) -> bool {
-    fields.sections
-        || fields.tasks
-        || fields.links
-        || fields.title
-        || has_content_search
-        || has_task_filter
-        || has_section_filter
-}
+use build::{TitleMatcher, extract_title, matches_task_filter};
+use sort::{apply_sort, presort_index_entries};
 
 /// Find files matching the given filters and return them as a JSON array.
 ///
@@ -577,183 +445,6 @@ pub fn find(
         format,
         &json_output,
     )))
-}
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-/// Extract the title value for `--fields title`.
-///
-/// Priority:
-/// 1. `title` frontmatter property (if it is a string)
-/// 2. First H1 heading in the document outline
-/// 3. `serde_json::Value::Null` if neither found
-fn extract_title(
-    props: &indexmap::IndexMap<String, serde_json::Value>,
-    outline_sections: Option<&[OutlineSection]>,
-) -> serde_json::Value {
-    // 1. Frontmatter title property
-    if let Some(serde_json::Value::String(s)) = props.get("title") {
-        return serde_json::Value::String(s.clone());
-    }
-    // 2. First H1 heading from outline
-    if let Some(sections) = outline_sections {
-        for sec in sections {
-            if sec.level == 1
-                && let Some(ref heading) = sec.heading
-            {
-                return serde_json::Value::String(heading.clone());
-            }
-        }
-    }
-    serde_json::Value::Null
-}
-
-/// Pre-compiled title filter — avoids per-file regex compilation and repeated
-/// `to_lowercase()` allocation.
-enum TitleMatcher {
-    /// Case-insensitive substring: stores the lowered pattern.
-    Substring(String),
-    /// Pre-compiled case-insensitive regex.
-    Regex(regex::Regex),
-}
-
-impl TitleMatcher {
-    /// Parse a `--title` value into a compiled matcher.
-    ///
-    /// Returns `Err(CommandOutcome::UserError(...))` on invalid regex.
-    fn parse(pattern: &str) -> Result<Self, CommandOutcome> {
-        if let Some(regex_pat) = pattern.strip_prefix("~=") {
-            let effective = format!("(?i){regex_pat}");
-            match regex::RegexBuilder::new(&effective)
-                .size_limit(1 << 20)
-                .build()
-            {
-                Ok(re) => Ok(Self::Regex(re)),
-                Err(e) => Err(CommandOutcome::UserError(format!(
-                    "invalid --title regex: {regex_pat}\n{e}"
-                ))),
-            }
-        } else {
-            Ok(Self::Substring(pattern.to_lowercase()))
-        }
-    }
-
-    /// Returns true if the title value matches. `Null` titles never match.
-    fn matches(&self, title: &serde_json::Value) -> bool {
-        let title_str = match title {
-            serde_json::Value::String(s) => s.as_str(),
-            _ => return false,
-        };
-        match self {
-            Self::Substring(lowered) => title_str.to_lowercase().contains(lowered.as_str()),
-            Self::Regex(re) => re.is_match(title_str),
-        }
-    }
-}
-
-/// Return true if `tasks` satisfy `filter`.
-fn matches_task_filter(tasks: &[FindTaskInfo], filter: &FindTaskFilter) -> bool {
-    match filter {
-        FindTaskFilter::Any => !tasks.is_empty(),
-        FindTaskFilter::Todo => tasks.iter().any(|t| !t.done),
-        FindTaskFilter::Done => tasks.iter().any(|t| t.done),
-        FindTaskFilter::Status(c) => tasks.iter().any(|t| t.status.starts_with(*c)),
-    }
-}
-
-/// Apply the requested sort order to the results.
-fn apply_sort(
-    results: &mut [FileObject],
-    sort: Option<&SortField>,
-    link_graph: Option<&LinkGraph>,
-) {
-    match sort.unwrap_or(&SortField::File) {
-        SortField::File => results.sort_by(|a, b| a.file.cmp(&b.file)),
-        SortField::Modified => results.sort_by(|a, b| a.modified.cmp(&b.modified)),
-        SortField::BacklinksCount => {
-            results.sort_by(|a, b| {
-                let a_count = a.backlinks.as_ref().map_or_else(
-                    || link_graph.map_or(0, |g| g.backlinks(&a.file).len()),
-                    Vec::len,
-                );
-                let b_count = b.backlinks.as_ref().map_or_else(
-                    || link_graph.map_or(0, |g| g.backlinks(&b.file).len()),
-                    Vec::len,
-                );
-                b_count.cmp(&a_count)
-            });
-        }
-        SortField::LinksCount => {
-            results.sort_by(|a, b| {
-                let a_count = a.links.as_ref().map_or(0, Vec::len);
-                let b_count = b.links.as_ref().map_or(0, Vec::len);
-                b_count.cmp(&a_count)
-            });
-        }
-        SortField::Title => {
-            results.sort_by(|a, b| {
-                let a_val = a.title.as_ref();
-                let b_val = b.title.as_ref();
-                filter::compare_property_values(a_val, b_val).then_with(|| a.file.cmp(&b.file))
-            });
-        }
-        SortField::Property(key) => {
-            results.sort_by(|a, b| {
-                let a_val = a.properties.as_ref().and_then(|p| p.get(key));
-                let b_val = b.properties.as_ref().and_then(|p| p.get(key));
-                filter::compare_property_values(a_val, b_val).then_with(|| a.file.cmp(&b.file))
-            });
-        }
-    }
-}
-
-/// Pre-sort index entries by the requested sort key so that the early-exit
-/// optimisation can collect the first N matches in final order.
-///
-/// This mirrors `apply_sort` but operates on `&IndexEntry` references
-/// instead of `FileObject` values, avoiding construction of the full object.
-fn presort_index_entries(
-    entries: &mut [&IndexEntry],
-    sort: Option<&SortField>,
-    link_graph: &LinkGraph,
-) {
-    match sort.unwrap_or(&SortField::File) {
-        SortField::File => entries.sort_by(|a, b| a.rel_path.cmp(&b.rel_path)),
-        SortField::Modified => entries.sort_by(|a, b| a.modified.cmp(&b.modified)),
-        SortField::BacklinksCount => {
-            // Descending by backlink count — matches apply_sort.
-            entries.sort_by(|a, b| {
-                let a_count = link_graph.backlinks(&a.rel_path).len();
-                let b_count = link_graph.backlinks(&b.rel_path).len();
-                b_count.cmp(&a_count)
-            });
-        }
-        SortField::LinksCount => {
-            entries.sort_by(|a, b| {
-                let a_count = a.links.len();
-                let b_count = b.links.len();
-                b_count.cmp(&a_count)
-            });
-        }
-        SortField::Title => {
-            entries.sort_by(|a, b| {
-                let a_val = extract_title(&a.properties, Some(&a.sections));
-                let b_val = extract_title(&b.properties, Some(&b.sections));
-                filter::compare_property_values(Some(&a_val), Some(&b_val))
-                    .then_with(|| a.rel_path.cmp(&b.rel_path))
-            });
-        }
-        SortField::Property(key) => {
-            entries.sort_by(|a, b| {
-                let a_val = a.properties.get(key.as_str());
-                let b_val = b.properties.get(key.as_str());
-                filter::compare_property_values(a_val, b_val)
-                    .then_with(|| a.rel_path.cmp(&b.rel_path))
-            });
-        }
-    }
 }
 
 // ---------------------------------------------------------------------------
