+++ crates/hyalo-cli/src/commands/mod.rs
@@ -6,6 +6,7 @@ pub mod drop_index;
 pub mod find;
 pub mod init;
 pub mod links;
+pub mod mutation;
 pub mod mv;
 pub mod properties;
 pub mod read;
