+++ crates/hyalo-cli/src/lib.rs
@@ -1,6 +1,10 @@
+mod cli;
 pub mod commands;
 pub mod config;
 pub mod hints;
 pub mod output;
+mod run;
 pub mod suggest;
 pub mod warn;
+
+pub use run::run;
